chore: initial work on benchmarks
